+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Main - not released yet
 * Added unique ids to sensors so that they can be added to areas
+* Changed device class to power and unit to kW of sensor `kostal_generator` and `kostal_inverter`
 
 # V1.10
 Fixed usage of deprecated units to be compatible with HA Core 2025.1